requirements.txt
akinator.py==4.0.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
pyTelegramBotAPI==3.7.3
requests==2.24.0
six==1.15.0
urllib3==1.25.10
